requirements.txt
pygame>=2.5.0
pyserial>=3.5
opencv-python>=4.8.0
Pillow>=10.0.0
